Add support for a client to connect with multiple end-points.
Modifies the -H parameter to take in a list of end-points to
connect to when in client mode to facilitate testing setups such
as multiple VMs on a server.  The -H option now takes in a comma
delimited list of IPs and ports in the following format:
<ip>:<control port>/<data port>,...

Code has similarly been changed to support this mode of operation,
threads on the client round robin assigning flows specified in -F
to each end-point.

Signed-off-by: Geoffrey Blake (Geoffrey.Blake@arm.com)