peer mode: Add ability to slave and coordinate multiple clients together udp_flood mode: Add fixes and enhancements to udp flood mode
This patch adds the concept of slave clients and a master control client
for testing servers from multiple machines to allow for scaling up of
traffic and flows.  This adds three new parameters to the benchmark
applications:  --slave-mode tells the client that it is going to wait
for a master to coordinate with it on when to go, -g which allocates
an additional port for peers to communicate on, and --slaves which is a
list of hosts and ports for the master node to use to communicate with
all the slaves and provide the distributed barrier on when to start
testing.

Peering mode adds a dependency with the libczmq library for talking
between peers.

This patch also adds enhancements to UDP flood mode, such as doing
one-shot mode on epoll to try and reduce unfairness between sockets
and also try and avoid dropped packets as much as possible.

Signed-off-by: Geoffrey Blake (Geoffrey.Blake@arm.com)