Bug-fixes: Fix buffer overrun and improper czmq usage
Patch fixes buffer overrun in common.c if there is a long host list
being used, simple fix for now is to increase the statically allocated
buffer size.  May need to revisit and make it more robust in the future.

Fixes in control_plane.c for improper use of zmsg_t with zmsg_send and
zmsg_recv that led to losing messages.  Testing with more than 2 peers
now works successfully.

Signed-off-by: Geoffrey Blake (Geoffrey.Blake@arm.com)